+++ starkbank/user/__project.py
@@ -3,11 +3,12 @@ from .__user import User
 
 class Project(User):
     """# Project object
-    The Project object is the main authentication entity for the SDK.
-    All requests to the Stark Bank API must be authenticated via a project,
+    The Project object is an authentication entity for the SDK that is permanently
+    linked to a specific Workspace.
+    All requests to the Stark Bank API must be authenticated via an SDK user,
     which must have been previously created at the Stark Bank website
     [https://sandbox.web.starkbank.com] or [https://web.starkbank.com]
-    before you can use it in this SDK. Projects may be passed as a parameter on
+    before you can use it in this SDK. Projects may be passed as the user parameter on
     each request or may be defined as the default user at the start (See README).
     ## Parameters (required):
     - id [string]: unique id required to identify project. ex: "5656565656565656"
@@ -29,3 +30,6 @@ class Project(User):
             private_key=private_key,
             environment=environment,
         )
+
+    def access_id(self):
+        return "project/{id}".format(id=self.id)